Hello,
This was my first IOS app that I am launching to the app store. It is called CountryFacts and what this does is that it teaches users what country matches the capital and the flag. I feel like we weren't properly taught about world geography during school so I developed an app where you can learn quickly and easily.